(PC-16477) feat(profile): add BeneficiaryCeilingsNew and files reorganization"
