+++ src/features/cheatcodes/pages/AppComponents/AppComponents.tsx
@@ -14,9 +14,9 @@ import { CulturalSurveyProgressBar } from 'features/culturalSurvey/components/Cu
 import { navigateToHome } from 'features/navigation/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
 import { BeneficiaryCeilings } from 'features/profile/components/BeneficiaryCeilings'
-import { BeneficiaryHeaderNew } from 'features/profile/components/BeneficiaryHeaderNew'
+import { SubscriptionMessageBadge } from 'features/profile/components/headers/Badges/SubscriptionMessageBadge'
+import { BeneficiaryHeaderNew } from 'features/profile/components/headers/BeneficiaryHeader/BeneficiaryHeaderNew'
 import { NonBeneficiaryHeader } from 'features/profile/components/NonBeneficiaryHeader'
-import { SubscriptionMessageBadge } from 'features/profile/components/SubscriptionMessageBadge'
 import { SelectionLabel } from 'features/search/atoms/SelectionLabel'
 import { formatToSlashedFrenchDate } from 'libs/dates'
 import { MAP_CATEGORY_ID_TO_ICON } from 'libs/parsers'
