+++ src/features/profile/components/NonBeneficiaryHeader.tsx
@@ -7,9 +7,9 @@ import { SubscriptionMessage } from 'api/gen'
 import { useDepositAmountsByAge } from 'features/auth/api'
 import { useBeneficiaryValidationNavigation } from 'features/auth/signup/useBeneficiaryValidationNavigation'
 import { useNextSubscriptionStep } from 'features/auth/signup/useNextSubscriptionStep'
-import { IdentityCheckPendingBadge } from 'features/profile/components/IdentityCheckPendingBadge'
-import { SubscriptionMessageBadge } from 'features/profile/components/SubscriptionMessageBadge'
-import { YoungerBadge } from 'features/profile/components/YoungerBadge'
+import { IdentityCheckPendingBadge } from 'features/profile/components/headers/Badges/IdentityCheckPendingBadge'
+import { SubscriptionMessageBadge } from 'features/profile/components/headers/Badges/SubscriptionMessageBadge'
+import { YoungerBadge } from 'features/profile/components/headers/Badges/YoungerBadge'
 import { useIsUserUnderage } from 'features/profile/utils'
 import { formatToSlashedFrenchDate } from 'libs/dates'
 import { PageHeader } from 'ui/components/headers/PageHeader'
