+++ src/features/profile/components/headers/Badges/IdentityCheckPendingBadge.tsx
@@ -1,7 +1,7 @@
 import { t } from '@lingui/macro'
 import React from 'react'
 
-import { ProfileBadge } from 'features/profile/components/ProfileBadge'
+import { ProfileBadge } from 'features/profile/components/headers/Badges/ProfileBadge'
 import { Clock } from 'ui/svg/icons/Clock'
 
 export const IdentityCheckPendingBadge = () => (
