+++ src/features/profile/pages/Profile.tsx
@@ -10,7 +10,7 @@ import { v4 as uuidv4 } from 'uuid'
 import { useAuthContext, useLogoutRoutine } from 'features/auth/AuthContext'
 import { useFavoritesState } from 'features/favorites/pages/FavoritesWrapper'
 import { useUserProfileInfo } from 'features/profile/api'
-import { ProfileHeader } from 'features/profile/components/ProfileHeader'
+import { ProfileHeader } from 'features/profile/components/headers/ProfileHeader/ProfileHeader'
 import { ProfileContainer } from 'features/profile/components/reusables'
 import { SectionWithSwitch } from 'features/profile/components/SectionWithSwitch'
 import { env } from 'libs/environment'
